backend/requirements.txt
Django>=4.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.2
Pillow>=10.0.0
drf-spectacular>=0.26.4
django-q2>=1.6.0
whitenoise>=6.5.0
gunicorn>=21.2.0
